Reformat Main class.
